requirements.txt
discord.py
pymysql
matplotlib
python-dotenv